% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_installed}
\alias{cv_installed}
\title{Detect OpenCV Installation}
\usage{
cv_installed()
}
\value{
A logical indicating the presence or absence of OpenCV.
}
\description{
This function detects whether OpenCV is correctly installed on
the system.
}
\seealso{
\code{\link[=install_cv]{install_cv()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
