% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_version}
\alias{cv_version}
\title{Detect OpenCV Version}
\usage{
cv_version()
}
\value{
A character string. NA indicates that OpenCV is not installed on the
system.
}
\description{
This function detects the version of OpenCV installed on the
system.
}
\seealso{
\code{\link[=install_cv]{install_cv()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
