% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fourcc}
\alias{fourcc}
\title{Codec Name to FOURCC Code}
\usage{
fourcc(str)
}
\arguments{
\item{x}{A 4-element character chain corresponding to the name of a valid
video codec. A list of valid codec names can be found at this archived
page of the fourcc site
\href{https://www.fourcc.org/codecs.php}{https://www.fourcc.org/codecs.php}.}
}
\value{
An integer value corresponding to the FOURCC code of the video codec.
}
\description{
\code{fourcc} translates the 4-character name of a video codec
into its corresponding \href{https://www.fourcc.org/codecs.php}{FOURCC}
code.
}
\examples{
fourcc("xvid")

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
