% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fps}
\alias{fps}
\title{Framerate of a Python VideoCapture Object}
\usage{
fps(x)
}
\arguments{
\item{x}{A Python VideoCapture object.}
}
\value{
A numeric value.
}
\description{
This function returns the framerate of a Python VideoCapture
object.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
