% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{install_cv}
\alias{install_cv}
\title{Install and Update OpenCV}
\usage{
install_cv(python_version = "3.12.5")
}
\arguments{
\item{python_version}{A character string indicating the version of Python you
would like OpenCV to run on (default: "3.12.5"). Not all versions of Python
will necessarily work on your system, but the chosen default works on most
systems that we tested so far.}
}
\value{
If the installation/update completes successfully, a data frame
indicating the location of the OpenCV installation and its version number.
}
\description{
This function automates the installation/updating of OpenCV and
all its Python dependencies in a dedicated Python virtual environment for
use with the \link{trackRcv} apps.
}
\note{
If the requested version of Python is not activated on your system, this
function will attempt to install it first before creating the dedicated
Python virtual environment.
}
\seealso{
\code{\link[=remove_cv]{remove_cv()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
