% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{is_image}
\alias{is_image}
\title{Test for a Numpy Array}
\usage{
is_image(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A logical indicating whether the object is a Numpy array (TRUE) or
not (FALSE).
}
\description{
This function tests whether an object is a Numpy array.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
