% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{is_video_capture}
\alias{is_video_capture}
\title{Test for a Python VideoCapture Object}
\usage{
is_video_capture(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A logical indicating whether the object is a Python VideoCapture
object (TRUE) or not (FALSE).
}
\description{
This function tests whether an object is a Python VideoCapture
object.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
