% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{n_col}
\alias{n_col}
\title{The Number of Columns in a Numpy Array}
\usage{
n_col(x)
}
\arguments{
\item{x}{A Numpy array.}
}
\value{
A numeric value.
}
\description{
This function returns the number of columns present in a Numpy
array.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
