% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{n_frames}
\alias{n_frames}
\title{The Number of Frames in a Python VideoCapture Object}
\usage{
n_frames(x)
}
\arguments{
\item{x}{A Python VideoCapture object.}
}
\value{
A numeric value.
}
\description{
This function returns the number of frames present in a Python
VideoCapture object.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
