% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{n_row}
\alias{n_row}
\title{The Number of Rows in a Numpy Array}
\usage{
n_row(x)
}
\arguments{
\item{x}{A Numpy array.}
}
\value{
A numeric value.
}
\description{
This function returns the number of rows present in a Numpy
array.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
