% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{read_frame}
\alias{read_frame}
\title{Read Specific Frame in a Python VideoCapture Object}
\usage{
read_frame(x, i)
}
\arguments{
\item{x}{A Python VideoCapture object.}

\item{i}{The 1-indexed frame to be read.}
}
\value{
A Numpy array.
}
\description{
This function reads a specific frame of a Python VideoCapture
object and returns it as Numpy array.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
