% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{remove_cv}
\alias{remove_cv}
\title{Remove OpenCV}
\usage{
remove_cv()
}
\value{
Nothing.
}
\description{
This function automates the removal of OpenCV and all its Python
dependencies from your system.
}
\note{
The function will only remove the dedicated Python virtual environment from
your system. If Python was installed during the execution of
\code{\link[=install_cv]{install_cv()}}, it will not be removed.
}
\seealso{
\code{\link[=install_cv]{install_cv()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
