% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{trackFixer}
\alias{trackFixer}
\title{Fix Tracking Errors}
\usage{
trackFixer(...)
}
\arguments{
\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the save
button is used in the app.
}
\description{
Shiny-based GUI to fix a posteriori common tracking errors, such
as removing unwanted tracks, fixing swapped track IDs and reconnecting
tracks.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
