% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{trackPlayer}
\alias{trackPlayer}
\title{Display and Export Video with Track Overlay}
\usage{
trackPlayer(...)
}
\arguments{
\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the export
button is used in the app.
}
\description{
Shiny-based GUI to display and export a video with tracks
overlaid on top.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
