% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{trackR}
\alias{trackR}
\title{App Launcher}
\usage{
trackR(...)
}
\arguments{
\item{...}{Parameters to be passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
This function does not return anything.
}
\description{
This function starts an app launcher that gives the user access
to all the Shiny apps provided with trackRcv.
}
\examples{
\dontrun{
trackR()
}

}
\seealso{
\code{\link[shiny:runApp]{shiny::runApp()}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
