% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{trackR}
\alias{trackR}
\title{Video Tracking App}
\usage{
trackR(...)
}
\arguments{
\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the
tracking is successfully completed.
}
\description{
A video tracking software with a Shiny-based GUI.
}
\examples{
# Run trackR in an external window
\dontrun{trackR()}

# Run trackR classic inside RStudio
\dontrun{trackR(launch.browser = rstudioapi::viewer)}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
